Refactored tests to follow page object pattern
Instead of repeating the same Playwright calls to fetch the locators we
use the Page Object Model pattern for all currently used pages.

I experimented with two approaches. The first approach was to reflect
all possible actions with dedicated functions in page object. This ended
up being a lot of boilerplate code. The second approach was to create
the most popular selectors and save them as page object fields. Those
locators that are used only one time, usually specific to the given
test, are not included in the page object but created manually. This
approach feels to be better for the current state of the project as it
allows to avoid re-creating the same locators over and over again and at
the same time we are not introducing too much boilerplate code.